Use ConvertAnsi pretty-format plugin as snapshotSerializer
