+++ package.json
@@ -1,6 +1,8 @@
 {
   "private": true,
   "devDependencies": {
+    "ansi-regex": "^2.0.0",
+    "ansi-styles": "^3.0.0",
     "babel-core": "^6.23.1",
     "babel-eslint": "^7.1.1",
     "babel-plugin-syntax-trailing-function-commas": "^6.13.0",
@@ -84,6 +86,9 @@
       "^.+\\.js$": "<rootDir>/packages/babel-jest"
     },
     "setupTestFrameworkScriptFile": "<rootDir>/testSetupFile.js",
+    "snapshotSerializers": [
+      "<rootDir>/packages/pretty-format/build/plugins/ConvertAnsi.js"
+    ],
     "testEnvironment": "./packages/jest-environment-node",
     "testPathIgnorePatterns": [
       "/node_modules/",
