+++ packages/jest-matchers/src/__tests__/matchers-test.js
@@ -13,63 +13,6 @@
 
 const jestExpect = require('../');
 const {stringify} = require('jest-matcher-utils');
-const {toMatchSnapshot} = require('jest-snapshot');
-
-const toHumanReadableAnsi = text => {
-  const style = require('ansi-styles');
-  const reg = require('ansi-regex');
-  return text.replace(reg(), (match, offset, string) => {
-    switch (match) {
-      case style.red.close:
-      case style.green.close:
-        return '</>';
-      case style.red.open:
-        return '<red>';
-      case style.green.open:
-        return '<green>';
-      default:
-        return '';
-    }
-  });
-};
-
-function toThrowErrorMatchingAnsiSnapshot(received: any, expected: void) {
-  this.dontThrow && this.dontThrow();
-  const {isNot} = this;
-
-  if (isNot) {
-    throw new Error(
-      'Jest: `.not` cannot be used with `.toThrowErrorMatchingAnsiSnapshot()`.',
-    );
-  }
-
-  this.utils.ensureNoExpected(expected, '.toThrowErrorMatchingAnsiSnapshot');
-
-  let error;
-
-  try {
-    received();
-  } catch (e) {
-    error = e;
-  }
-
-  if (error === undefined) {
-    throw new Error(
-      this.utils.matcherHint(
-        '.toThrowErrorMatchingAnsiSnapshot',
-        '() => {}',
-        ''
-      ) +
-        '\n\n' +
-        `Expected the function to throw an error.\n` +
-        `But it didn't throw anything.`
-    );
-  }
-
-  return toMatchSnapshot.call(this, toHumanReadableAnsi(error.message));
-}
-
-expect.extend({toThrowErrorMatchingAnsiSnapshot});
 
 describe('.toBe()', () => {
   it('does not throw', () => {
@@ -94,14 +37,14 @@ describe('.toBe()', () => {
   ].forEach(([a, b]) => {
     it(`fails for: ${stringify(a)} and ${stringify(b)}`, () => {
       expect(() => jestExpect(a).toBe(b))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
   [false, 1, 'a', undefined, null, {}, []].forEach(v => {
     it(`fails for '${stringify(v)}' with '.not'`, () => {
       expect(() => jestExpect(v).not.toBe(v))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -110,7 +53,7 @@ describe('.toBe()', () => {
     obj.circular = obj;
 
     expect(() => jestExpect(obj).toBe({}))
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
   });
 
   test('assertion error matcherResult property contains matcher name, expected and actual values', () => {
@@ -153,7 +96,7 @@ describe('.toEqual()', () => {
   ].forEach(([a, b]) => {
     test(`expect(${stringify(a)}).toEqual(${stringify(b)})`, () => {
       expect(() => jestExpect(a).toEqual(b))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -186,7 +129,7 @@ describe('.toEqual()', () => {
   ].forEach(([a, b]) => {
     test(`expect(${stringify(a)}).not.toEqual(${stringify(b)})`, () => {
       expect(() => jestExpect(a).not.toEqual(b))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -216,7 +159,7 @@ describe('.toBeInstanceOf()', () => {
   ].forEach(([a, b]) => {
     test(`passing ${stringify(a)} and ${stringify(b)}`, () => {
       expect(() => jestExpect(a).not.toBeInstanceOf(b))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
 
       jestExpect(a).toBeInstanceOf(b);
     });
@@ -231,7 +174,7 @@ describe('.toBeInstanceOf()', () => {
   ].forEach(([a, b]) => {
     test(`failing ${stringify(a)} and ${stringify(b)}`, () => {
       expect(() => jestExpect(a).toBeInstanceOf(b))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
 
       jestExpect(a).not.toBeInstanceOf(b);
     });
@@ -239,17 +182,17 @@ describe('.toBeInstanceOf()', () => {
 
   it('throws if constructor is not a function', () => {
     expect(() => jestExpect({}).toBeInstanceOf(4))
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
   });
 });
 
 describe('.toBeTruthy(), .toBeFalsy()', () => {
   it('does not accept arguments', () => {
     expect(() => jestExpect(0).toBeTruthy(null))
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
 
     expect(() => jestExpect(0).toBeFalsy(null))
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
   });
 
   [{}, [], true, 1, 'a', 0.5, new Map(), () => {}, Infinity].forEach(v => {
@@ -258,10 +201,10 @@ describe('.toBeTruthy(), .toBeFalsy()', () => {
       jestExpect(v).not.toBeFalsy();
 
       expect(() => jestExpect(v).not.toBeTruthy())
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
 
       expect(() => jestExpect(v).toBeFalsy())
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -271,10 +214,10 @@ describe('.toBeTruthy(), .toBeFalsy()', () => {
       jestExpect(v).not.toBeTruthy();
 
       expect(() => jestExpect(v).toBeTruthy())
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
 
       expect(() => jestExpect(v).not.toBeFalsy())
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 });
@@ -285,14 +228,14 @@ describe('.toBeNaN()', () => {
       jestExpect(v).toBeNaN();
 
       expect(() => jestExpect(v).not.toBeNaN())
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
   it('throws', () => {
     [1, '', null, undefined, {}, [], 0.2, 0, Infinity, -Infinity].forEach(v => {
       expect(() => jestExpect(v).toBeNaN())
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
 
       jestExpect(v).not.toBeNaN();
     });
@@ -305,7 +248,7 @@ describe('.toBeNull()', () => {
       jestExpect(v).not.toBeNull();
 
       expect(() => jestExpect(v).toBeNull())
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -313,7 +256,7 @@ describe('.toBeNull()', () => {
     jestExpect(null).toBeNull();
 
     expect(() => jestExpect(null).not.toBeNull())
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
   });
 });
 
@@ -324,10 +267,10 @@ describe('.toBeDefined(), .toBeUndefined()', () => {
       jestExpect(v).not.toBeUndefined();
 
       expect(() => jestExpect(v).not.toBeDefined())
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
 
       expect(() => jestExpect(v).toBeUndefined())
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -336,10 +279,10 @@ describe('.toBeDefined(), .toBeUndefined()', () => {
     jestExpect(undefined).not.toBeDefined();
 
     expect(() => jestExpect(undefined).toBeDefined())
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
 
     expect(() => jestExpect(undefined).not.toBeUndefined())
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
   });
 });
 
@@ -370,28 +313,28 @@ describe(
 
       it(`throws: [${small}, ${big}]`, () => {
         expect(() => jestExpect(small).toBeGreaterThan(big))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
 
         expect(() => jestExpect(small).not.toBeLessThan(big))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
 
         expect(() => jestExpect(big).not.toBeGreaterThan(small))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
 
         expect(() => jestExpect(big).toBeLessThan(small))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
 
         expect(() => jestExpect(small).toBeGreaterThanOrEqual(big))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
 
         expect(() => jestExpect(small).not.toBeLessThanOrEqual(big))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
 
         expect(() => jestExpect(big).not.toBeGreaterThanOrEqual(small))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
 
         expect(() => jestExpect(big).toBeLessThanOrEqual(small))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
       });
     });
 
@@ -407,10 +350,10 @@ describe(
         jestExpect(n1).toBeLessThanOrEqual(n2);
 
         expect(() => jestExpect(n1).not.toBeGreaterThanOrEqual(n2))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
 
         expect(() => jestExpect(n1).not.toBeLessThanOrEqual(n2))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
       });
     });
   },
@@ -456,7 +399,7 @@ describe('.toContain(), .toContainEqual()', () => {
       jestExpect(list).toContain(v);
 
       expect(() => jestExpect(list).not.toContain(v))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -470,13 +413,13 @@ describe('.toContain(), .toContainEqual()', () => {
       jestExpect(list).not.toContain(v);
 
       expect(() => jestExpect(list).toContain(v))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
   test('error cases', () => {
     expect(() => jestExpect(null).toContain(1))
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
   });
 
   [
@@ -492,7 +435,7 @@ describe('.toContain(), .toContainEqual()', () => {
     it(`'${stringify(list)}' contains a value equal to '${stringify(v)}'`, () => {
       jestExpect(list).toContainEqual(v);
       expect(() => jestExpect(list).not.toContainEqual(v))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -503,13 +446,13 @@ describe('.toContain(), .toContainEqual()', () => {
       jestExpect(list).not.toContainEqual(v);
 
       expect(() => jestExpect(list).toContainEqual(v))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
   test('error cases for toContainEqual', () => {
     expect(() => jestExpect(null).toContainEqual(1))
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
   });
 });
 
@@ -526,7 +469,7 @@ describe('.toBeCloseTo()', () => {
       jestExpect(n1).toBeCloseTo(n2);
 
       expect(() => jestExpect(n1).not.toBeCloseTo(n2))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -537,7 +480,7 @@ describe('.toBeCloseTo()', () => {
   ].forEach(([n1, n2]) => {
     it(`throws: [${n1}, ${n2}]`, () => {
       expect(() => jestExpect(n1).toBeCloseTo(n2))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
 
       jestExpect(n1).not.toBeCloseTo(n2);
     });
@@ -552,7 +495,7 @@ describe('.toBeCloseTo()', () => {
       jestExpect(n1).toBeCloseTo(n2, p);
 
       expect(() => jestExpect(n1).not.toBeCloseTo(n2, p))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 });
@@ -563,14 +506,14 @@ describe('.toMatch()', () => {
       jestExpect(n1).toMatch(n2);
 
       expect(() => jestExpect(n1).not.toMatch(n2))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
   [['bar', 'foo'], ['bar', /foo/]].forEach(([n1, n2]) => {
     it(`throws: [${n1}, ${n2}]`, () => {
       expect(() => jestExpect(n1).toMatch(n2))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -588,7 +531,7 @@ describe('.toMatch()', () => {
         ` [${stringify(n1)}, ${stringify(n2)}]`,
       () => {
         expect(() => jestExpect(n1).toMatch(n2))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
       },
     );
   });
@@ -606,7 +549,7 @@ describe('.toMatch()', () => {
         ` [${stringify(n1)}, ${stringify(n2)}]`,
       () => {
         expect(() => jestExpect(n1).toMatch(n2))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
       },
     );
   });
@@ -627,7 +570,7 @@ describe('.toHaveLength', () => {
     test(`expect(${stringify(received)}).toHaveLength(${length})`, () => {
       jestExpect(received).toHaveLength(length);
       expect(() => jestExpect(received).not.toHaveLength(length))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -641,17 +584,17 @@ describe('.toHaveLength', () => {
     test(`expect(${stringify(received)}).toHaveLength(${length})`, () => {
       jestExpect(received).not.toHaveLength(length);
       expect(() => jestExpect(received).toHaveLength(length))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
   test('error cases', () => {
     expect(() => jestExpect({a: 9}).toHaveLength(1))
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
     expect(() => jestExpect(0).toHaveLength(1))
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
     expect(() => jestExpect(undefined).toHaveLength(1))
-      .toThrowErrorMatchingAnsiSnapshot();
+      .toThrowErrorMatchingSnapshot();
   });
 });
 
@@ -668,7 +611,7 @@ describe('.toHaveProperty()', () => {
       () => {
         jestExpect(obj).toHaveProperty(keyPath, value);
         expect(() => jestExpect(obj).not.toHaveProperty(keyPath, value))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
       },
     );
   });
@@ -688,7 +631,7 @@ describe('.toHaveProperty()', () => {
       `{pass: false} expect(${stringify(obj)}).toHaveProperty('${keyPath}', ${stringify(value)})`,
       () => {
         expect(() => jestExpect(obj).toHaveProperty(keyPath, value))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
         jestExpect(obj).not.toHaveProperty(keyPath, value);
       },
     );
@@ -704,7 +647,7 @@ describe('.toHaveProperty()', () => {
       () => {
         jestExpect(obj).toHaveProperty(keyPath);
         expect(() => jestExpect(obj).not.toHaveProperty(keyPath))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
       },
     );
   });
@@ -720,7 +663,7 @@ describe('.toHaveProperty()', () => {
       `{pass: false} expect(${stringify(obj)}).toHaveProperty('${keyPath}')`,
       () => {
         expect(() => jestExpect(obj).toHaveProperty(keyPath))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
         jestExpect(obj).not.toHaveProperty(keyPath);
       },
     );
@@ -737,7 +680,7 @@ describe('.toHaveProperty()', () => {
       `{error} expect(${stringify(obj)}).toHaveProperty('${keyPath}')`,
       () => {
         expect(() => jestExpect(obj).toHaveProperty(keyPath))
-          .toThrowErrorMatchingAnsiSnapshot();
+          .toThrowErrorMatchingSnapshot();
       },
     );
   });
@@ -766,7 +709,7 @@ describe('toMatchObject()', () => {
     it(`{pass: true} expect(${stringify(n1)}).toMatchObject(${stringify(n2)})`, () => {
       jestExpect(n1).toMatchObject(n2);
       expect(() => jestExpect(n1).not.toMatchObject(n2))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -798,7 +741,7 @@ describe('toMatchObject()', () => {
     it(`{pass: false} expect(${stringify(n1)}).toMatchObject(${stringify(n2)})`, () => {
       jestExpect(n1).not.toMatchObject(n2);
       expect(() => jestExpect(n1).toMatchObject(n2))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 
@@ -816,7 +759,7 @@ describe('toMatchObject()', () => {
   ].forEach(([n1, n2]) => {
     it(`throws expect(${stringify(n1)}).toMatchObject(${stringify(n2)})`, () => {
       expect(() => jestExpect(n1).toMatchObject(n2))
-        .toThrowErrorMatchingAnsiSnapshot();
+        .toThrowErrorMatchingSnapshot();
     });
   });
 });
